+++ contracts/test/UniswapV3PairEchidnaTest.sol
@@ -39,7 +39,7 @@ contract UniswapV3PairEchidnaTest {
         pair = UniswapV3Pair(factory.createPair(address(token0), address(token1), fee));
     }
 
-    function initializePair(int16 tick) public {
+    function initializePair(int24 tick) public {
         pair.initialize(tick);
     }
 
@@ -54,8 +54,8 @@ contract UniswapV3PairEchidnaTest {
     }
 
     function setPosition(
-        int16 tickLower,
-        int16 tickUpper,
+        int24 tickLower,
+        int24 tickUpper,
         int128 liquidityDelta
     ) external {
         pair.setPosition(tickLower, tickUpper, liquidityDelta);
@@ -78,12 +78,12 @@ contract UniswapV3PairEchidnaTest {
     }
 
     function echidna_tickIsWithinBounds() external view returns (bool) {
-        int16 tick = pair.tickCurrent();
+        int24 tick = pair.tickCurrent();
         return (tick < TickMath.MAX_TICK && tick >= TickMath.MIN_TICK);
     }
 
     function echidna_priceIsWithinTickCurrent() external view returns (bool) {
-        int16 tick = pair.tickCurrent();
+        int24 tick = pair.tickCurrent();
         FixedPoint128.uq128x128 memory priceCurrent = FixedPoint128.uq128x128(pair.priceCurrent());
         return (TickMath.getRatioAtTick(tick)._x <= priceCurrent._x &&
             TickMath.getRatioAtTick(tick + 1)._x > priceCurrent._x);
